Y-axis force converted to double

+++ bolt_bearing_355/temp_function.h
@@ -127,7 +127,7 @@ void enter_thick_info(WINDOW *a, int color_pair, int num_arr, char *arr, char ch
 double enter_force_x(WINDOW *sub1, WINDOW *a, int color_pair);
 
 // Ввод силы по оси Y
-unsigned int enter_force_y(WINDOW *sub1, WINDOW *a, int color_pair);
+double enter_force_y(WINDOW *sub1, WINDOW *a, int color_pair);
 
 // Ввод силы по оси Z
 unsigned int enter_force_z(WINDOW *sub1, WINDOW *a, WINDOW *b, int color_pair);
@@ -157,7 +157,7 @@ double calc_tens_n_bt(unsigned int r_bt, double a_bn);
 void draw_coord_sys(WINDOW *b);
 
 // Рисуем силы на системе координат
-void draw_force_coord_sys(WINDOW *b, double force_x, unsigned int force_y, unsigned int force_z);
+void draw_force_coord_sys(WINDOW *b, double force_x, double force_y, unsigned int force_z);
 
 // Вывод результатов расчета_1
 void output_results_1(WINDOW *sub1, double max_sher_result, double max_bear_result, double max_tens_result);
